kernel/fde,secboot: move infra to run fde-reveal-key to kernel/fde
deprecate KeyName in fde-reveal-key "reveal"
