+++ secboot/export_test.go
@@ -22,7 +22,6 @@ package secboot
 
 import (
 	"io"
-	"time"
 
 	sb "github.com/snapcore/secboot"
 )
@@ -185,29 +184,5 @@ func MockFDEHasRevealKey(f func() bool) (restore func()) {
 	fdeHasRevealKey = f
 	return func() {
 		fdeHasRevealKey = old
-	}
-}
-
-func MockFdeRevealKeyCommandExtra(args []string) (restore func()) {
-	oldFdeRevealKeyCommandExtra := fdeRevealKeyCommandExtra
-	fdeRevealKeyCommandExtra = args
-	return func() {
-		fdeRevealKeyCommandExtra = oldFdeRevealKeyCommandExtra
-	}
-}
-
-func MockFdeRevealKeyRuntimeMax(d time.Duration) (restore func()) {
-	oldFdeRevealKeyRuntimeMax := fdeRevealKeyRuntimeMax
-	fdeRevealKeyRuntimeMax = d
-	return func() {
-		fdeRevealKeyRuntimeMax = oldFdeRevealKeyRuntimeMax
-	}
-}
-
-func MockFdeRevealKeyPollWaitParanoiaFactor(n int) (restore func()) {
-	oldFdeRevealKeyPollWaitParanoiaFactor := fdeRevealKeyPollWaitParanoiaFactor
-	fdeRevealKeyPollWaitParanoiaFactor = n
-	return func() {
-		fdeRevealKeyPollWaitParanoiaFactor = oldFdeRevealKeyPollWaitParanoiaFactor
 	}
 }
